+++ src/spUtils.ts
@@ -1,4 +1,5 @@
 ﻿import {
+  CompletionItemKind,
   Location,
   MarkdownString,
   Range,
@@ -7,6 +8,9 @@
 import { URI } from "vscode-uri";
 import { existsSync, lstatSync } from "fs";
 import { resolve, extname } from "path";
+import { SPItem } from "./Backend/Items/spItems";
+import { MethodItem } from "./Backend/Items/spMethodItem";
+import { PropertyItem } from "./Backend/Items/spPropertyItem";
 
 /**
  * Parse a Sourcemod JSDoc documentation string and convert it to a MarkdownString.
@@ -90,4 +94,33 @@ export function checkMainPath(mainPath: string): boolean {
  */
 export function locationFromRange(filePath: string, range: Range): Location {
   return new Location(URI.file(filePath), range);
+}
+
+/**
+ * Check if an item should be a constructor instead of a methodmap, given the line it was found in.
+ * If it is a constructor, return the correct item. Otherwise, return the original object.
+ * @param  {SPItem} item  The item to check.
+ * @param  {Range} range  The range of the match.
+ * @param  {Map<string, MethodItem|PropertyItem>} methodsAndProperties  The methodsAndProperties of the parser.
+ * @param  {string} line  The line at which the item was matched at.
+ * @returns SPItem
+ */
+export function checkIfConstructor(
+  item: SPItem,
+  range: Range,
+  methodsAndProperties: Map<string, MethodItem | PropertyItem>,
+  line: string
+): SPItem {
+  if (item.kind !== CompletionItemKind.Class) {
+    return item;
+  }
+  const subLine = line.substring(range.start.character);
+  const re = new RegExp(`^${item.name}\\s*\\(`);
+  if (!re.test(subLine)) {
+    return item;
+  }
+  return (
+    (methodsAndProperties.get(`${item.name}-${item.name}`) as MethodItem) ||
+    item
+  );
 }