+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added code folding for `#if` and `#endif` (see #210).
 - Added the possibility to use `${mainPath}` as the localRoot setting for the FTP/SFTP upload setting (see #209).
 - Separated primitive types from other types, like in the C/C++ extension.
+- Added better methodmap constructor support.
 
 ### Fixed